% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APSHP.R
\name{APKML2SHP}
\alias{APKML2SHP}
\title{AP Convert KLM to SHP}
\usage{
APKML2SHP(inFile, OSGeoPath = "C:\\\\OSGeo4W64", saveToFile = TRUE,
  readToMemory = FALSE, dropDescription = TRUE)
}
\arguments{
\item{inFile}{path to a KML file}

\item{OSGeoPath}{character. Path to the OSGeo4W installation directory.}

\item{saveToFile}{logical. Save object as a Shapefile. File will be saved with the same name and in the
same folder as the \code{inFile}.}

\item{readToMemory}{logical. Input will be return as Spatial object.}

\item{dropDescription}{logical. If set to TRUE, the Description field of the KML file will be dropped}
}
\description{
Read a KML file and save it to a SHP file and/or read it to memory. Requires the use of
'ogrinfo.exe'.
}
\examples{
\dontrun{
APKML2SHP("C:/Geodata/myfile.kml", dropDescription = FALSE)
}

}
