% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APopen.R
\name{APopen}
\alias{APopen}
\title{AP Folder Open}
\usage{
APopen(x)
}
\arguments{
\item{x}{character. File path. If path leads to a directory, it will open that directory. If it directs to a file, it
will open that file's directory}
}
\description{
Open a folder in Windows Explorer
}

