% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APpolyMerge.R
\name{APpolyMerge}
\alias{APpolyMerge}
\title{Merge polygons}
\usage{
APpolyMerge(polyList, newID = FALSE)
}
\arguments{
\item{polyList}{List. a List of SpatialPolygonsDataFrame objects}

\item{newID}{logical. If TRUE, the polygon IDs in \code{polyList} will be replaced to prevent duplicate IDs.}
}
\value{
A merged SpatialPolygonsDataFrame
}
\description{
Take a List of SpatialPolygonsDataFrame objects and merge them. This function automates the process of
assigning new polygon IDs, which is usually the issue that prevents merging.
}
