% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APprecise.R
\name{APprecise}
\alias{APprecise}
\title{AP Precision}
\usage{
APprecise(x, digits = 22)
}
\arguments{
\item{x}{numeric. Input value value}

\item{digits}{numeric. Number of digits to display}
}
\description{
Prints input value with a set number digits.
}

