% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APrasterfiles.R
\name{APrasterFiles}
\alias{APrasterFiles}
\title{AP Get Raster Files}
\usage{
APrasterFiles(filePath)
}
\arguments{
\item{filePath}{character. Path to file}
}
\description{
Get list of auxiliary raster files
}
\examples{
\dontrun{
APSHPfiles("C:/Geodata/myfile.shp")
}

}
