% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AProunder.R
\name{AProunder}
\alias{AProunder}
\title{AP Rounder}
\usage{
AProunder(value, interval, direction = "closest", snap = 0)
}
\arguments{
\item{value}{numeric input value value}

\item{interval}{numeric. The interval to which the input value should be rounded}

\item{direction}{character. The rounding direction. Can be \code{'closest'}, \code{'up'} or \code{'down'}}

\item{snap}{numeric. An origin value through which the interval with pass through. Default is 0.}
}
\value{
Rounded number
}
\description{
Provides extra options for rounding numbers, such as rounding a value to uneven intervals
and setting those intervals to 'snap' or pass through a defined origin value.
}

