% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APSHP.R
\name{APSHPdel}
\alias{APSHPdel}
\title{AP Delete Shapefile}
\usage{
APSHPdel(filePath)
}
\arguments{
\item{filePath}{character. Path to file}
}
\description{
Delete a Shapefile and all associated files
}
\examples{
\dontrun{
APSHPdel("C:/Geodata/myfile.shp")
}
}
