% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APSHP.R
\name{APSHPfiles}
\alias{APSHPfiles}
\title{AP Get Shapefile files}
\usage{
APSHPfiles(filePath)
}
\arguments{
\item{filePath}{character. Path to file}
}
\description{
Get all files associated to a shapefile
}
