% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APSHP.R
\name{APSHPread}
\alias{APSHPread}
\title{AP Read Shapefile}
\usage{
APSHPread(filePath, warnings = FALSE)
}
\arguments{
\item{filePath}{character. Path to file}

\item{warnings}{logical. If FALSE, then warnings will be suppressed}
}
\value{
SpatialPolygonsDataFrame
}
\description{
Read a Shapefile from a path
}
\examples{
\dontrun{
inPoly <- APSHPread("C:/Geodata/myfile.shp")
}

}
