% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APSHP.R
\name{APSHPsave}
\alias{APSHPsave}
\title{AP Save to SHP}
\usage{
APSHPsave(object, outfile, overwrite = FALSE)
}
\arguments{
\item{object}{a Spatial object}

\item{outfile}{path for file to be saved}

\item{overwrite}{logical. Allow function to overwrite existing file. If set to 'prompt', it will
ask user whether or to overwrite}
}
\description{
Save a Spatial type object to disk as a Shapefile.
}
\examples{
\dontrun{
APSHPsave(inPoly, outfile = "C:/Geodata/myfile.shp")
}

}
