% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APrasterfiles.R
\docType{data}
\name{rasterExtensions}
\alias{rasterExtensions}
\title{Raster Extensions}
\format{
list
}
\usage{
rasterExtensions
}
\description{
A list object, for which each element corresponds to a type of raster file. The elements
are character vectors of the extensions of the various metadata files that can be associated with
that type of raster.
}
\keyword{datasets}
