% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APSHP.R
\docType{data}
\name{SHPextensions}
\alias{SHPextensions}
\title{Shapefile extensions}
\format{
Character vector
}
\source{
https://en.wikipedia.org/wiki/Shapefile
}
\usage{
SHPextensions
}
\description{
A vector of extensions for the various file types associated with Esri Shapefiles.
}
\keyword{datasets}
