% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APcontours.R
\name{APcontours}
\alias{APcontours}
\title{AP Contours}
\usage{
APcontours(inRaster, interval)
}
\arguments{
\item{inRaster}{RasterLayer. A digital surface or digital elevation model}

\item{interval}{numeric. Interval for contour intervals}
}
\description{
Creates contours with rounded values
}

